+++ superset/migrations/versions/2024-12-06_15-19_8a23aa3bdd84_add_index_to_permission_view.py
@@ -44,5 +44,8 @@ def upgrade():
 
 
 def downgrade():
-    op.drop_index(op.f(perm_index_permission_id), table_name=perm_table)
-    op.drop_index(op.f(perm_index_view_menu), table_name=perm_table)
+    if table_has_index(perm_table, perm_index_permission_id):
+        op.drop_index(op.f(perm_index_permission_id), table_name=perm_table)
+
+    if table_has_index(perm_table, perm_index_view_menu):
+        op.drop_index(op.f(perm_index_view_menu), table_name=perm_table)
